requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
elevenlabs==1.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
pyttsx3==2.90
PyYAML==6.0.1
requests==2.32.3
setuptools==70.3.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
wheel==0.43.0
pydub
openai-whisper
deepgram-sdk
transformers
boto3